Use "at" property to define the start time of a repeating event.
This fixes the issue where setting a time in the past would set the initial trigger time at the current time. Thus, if it is 13:40 and I set a recurrent 30 minutes notification, I'd get a notification now and at 14:10, when I set the notification at 13:30 so it would notify me at 14:00.